Add cache buster to examples. This way the server is loaded each time allowing one to execute both example files in the same runner instance.
The buster is needed because each example file brings up and shuts down the server; when the
second test tries to bring up the server it gets the cased version of the server whcih
is already down. The server was brought up when the module was loaded so since we
have a cached module it won't come up again.