Add support for global setup. If your tests require a server be brought up instead of lifting and lowering it in each test files before and after hooks you can now define a special file called 'setup.js' and use the 'rootSuite' hook to define startup(before)/teardown(after) steps.
This way you can have this root suite added which will lift the server before running _any_ of
the tests then once all the test are done teardown the server.

I'm opting for convention over configuration here so you have to use the rootSuite hook
to denote this as your global setup/tear down and it _must_ be in a file called 'setup.js'
to ensure that its executed before any of the tests files. The later is about making
sure each test file's suite is a child of the root suite, we need to know about the
root suite first before generating each child suite.