:memo:Add notes on how to test dynamic routes
This change set adds a block in the Readme that explains how one would test
a route which is generated dynamically, such as a route with an ID that is
generated by a before hook.

This is how one would test a GET route for a collection where the before
hook is used to insert the resource into the collection to fetch in the
test.